# [Tizen] Enhances the Static Registrar