### Description of Change ###

This PR is an improvement over Static Registrar, which was introduced previous PR (#8642). We did various experiments to improve the startup performance of tizen application and provided many improvements by using Static Registrar. But we found some small improvements to the process of creating instances dynamically, and in this PR we're dealing with them.

As everyone knows, we are using expensive `Activator.CreateInstance()` to create a handler for a given view type. Yes, this is much more expensive than initializing with `new` operator. When using static registers, I thought about alternatives, and the experimental results were as follows. 
>_The test measured the time taken to create the instance of `CirclePageRendere` 10 times each in the Galaxy Watch._

- AS-IS. Using `Activator.CreateInstance()` -> **9.736 ms**

- case 1. Using  `new` operator -> **1.571 ms**
```cs
new CirclePageRenderer();
```
- case 2. Using Lambda expression with Action delegate -> **1.578 ms**
```cs
Action action = () => new CirclePageRenderer();
action.Invoke();
```

- case 3. Using Generic Template -> 3.81 ms
```cs
static T Create<T>() where T :new()
{
    return new T();
}
 
Create<CirclePageRenderer>();
```
- case 4. IL Emit -> 2.828 ms
```cs
public static class ILEmit
{
    public delegate object DynamicObjectActivator();
    private static readonly Dictionary<Type, DynamicObjectActivator> cache = new Dictionary<Type, DynamicObjectActivator>();
 
    public static object Build(Type t)
    {
        if (cache.TryGetValue(t, out DynamicObjectActivator value))
        {
            return value();
        }
 
        var dynamicMethod = new DynamicMethod("CreateInstance", t, Type.EmptyTypes, true);
        var ilGenerator = dynamicMethod.GetILGenerator();
        ilGenerator.Emit(OpCodes.Nop);
        ConstructorInfo emptyConstructor = t.GetConstructor(Type.EmptyTypes);
        ilGenerator.Emit(OpCodes.Newobj, emptyConstructor);
        ilGenerator.Emit(OpCodes.Ret);
 
        var del = (DynamicObjectActivator)dynamicMethod.CreateDelegate(typeof(DynamicObjectActivator));
        cache.Add(t, del);
        return del();
    }
}

ILEmit.Build(typeof(CirclePageRenderer);
```

As shown above test result, we can see that Case 1 (using Lambda expression with Action delegate) is as fast as `new` operator. So I applied this to Static Registrar.  If you want to register a custom handler in your app using the Static Registrar, you can do the following:

Before:
```cs
var option = new InitializationOptions(app, true);
var customRenderers = new Dictionary<Type, Type> { {typeof(CirclePage), typeof(CirclePageRenderer)} };
option.UseStaticRegistrar(StaticRegistrarStrategy.StaticRegistrarOnly, customRenderers, true);
Forms.Init(option);
```

After:
```cs
var option = new InitializationOptions(app, true);
var customRenderers = new Dictionary<Type,  Func<IRegisterable>> { {typeof(CirclePage), () => new CirclePageRenderer()} };
option.UseStaticRegistrar(StaticRegistrarStrategy.StaticRegistrarOnly, customRenderers, true);
Forms.Init(option);
```

### Issues Resolved ### 
None

### API Changes ###
Changed:
 - Dictionary<Type, Type> CustomHandlers => Dictionary<Type, Func<IRegisterable>> CustomHandlers

### Platforms Affected ### 
- Tizen

### Behavioral/Visual Changes ###
None

### Before/After Screenshots ### 
Not applicable

### PR Checklist ###

- [ ] Targets the correct branch
- [ ] Tests are passing (or failures are unrelated)